+++ src/domain/data.ts
@@ -1,6 +1,8 @@
 import { range } from "d3";
 
 import buildEnv from "src/env/build";
+import { OperatorObservationFieldsFragment } from "src/graphql/queries";
+import { weightedMean } from "src/utils/weighted-mean";
 
 export type ObservationValue = string | number | boolean | Date;
 export type GenericObservation = Record<string, ObservationValue>;
@@ -56,13 +58,32 @@ export const periods = range(
   -1
 ).map((d) => d.toString());
 
-export const priceComponents = [
+const allPriceComponents = [
   "total",
   "gridusage",
   "energy",
   "charge",
+  /**
+   * Metering costs are measured in CHF per year. For comparing between operators (on the map),
+   * this is the value that is used.
+   */
+  "annualmeteringcost",
+  /*
+   * For the details display, they are converted by Elcom to Rp/kWh to align with other price components.
+   */
+  "meteringrate",
+
+  /** We need to keep aidfee even if not shown on the map for the total to be OK */
   "aidfee",
-];
+] as const;
+
+export const mapPriceComponents = allPriceComponents.filter(
+  (x) => x !== "meteringrate" && x !== "aidfee"
+);
+export const detailsPriceComponents = allPriceComponents.filter(
+  (x) => x !== "annualmeteringcost"
+);
+
 export const products = ["cheapest", "standard"];
 export const categories = [
   "H1",
@@ -104,3 +125,13 @@ export const asElectricityCategory = (
 };
 
 export type ValueFormatter = (value: number) => string;
+
+export const getObservationsWeightedMean = (
+  obs: OperatorObservationFieldsFragment[]
+) => {
+  return weightedMean(
+    obs,
+    (d) => d.value ?? 0,
+    (d) => d.coverageRatio
+  );
+};