+++ src/env/client.ts
@@ -2,7 +2,7 @@ import z from "zod";
 
 const FlagSchema = z.array(z.string());
 
-const clientSchema = z.object({
+const clientBuildSchema = z.object({
   NEXT_PUBLIC_VERCEL_ENV: z.string().optional(),
   NEXT_PUBLIC_FLAGS: z
     .string()
@@ -16,11 +16,23 @@ const clientSchema = z.object({
         return [];
       }
     }),
+});
+
+export const clientBuildEnv = clientBuildSchema.parse({
+  NEXT_PUBLIC_VERCEL_ENV: process.env.NEXT_PUBLIC_VERCEL_ENV,
+  NEXT_PUBLIC_FLAGS: process.env.NEXT_PUBLIC_FLAGS,
+});
+
+// When adding a variable here, also add it in runtime-client-vars
+const clientRuntimeSchema = z.object({
   PUBLIC_URL: z.string(),
 });
 
-export default clientSchema.parse({
-  NEXT_PUBLIC_VERCEL_ENV: process.env.NEXT_PUBLIC_VERCEL_ENV,
-  NEXT_PUBLIC_FLAGS: process.env.NEXT_PUBLIC_FLAGS,
-  PUBLIC_URL: process.env.VERCEL_URL || process.env.PUBLIC_URL,
-});
+export const globalVariableName = `__RUNTIME_ENV__`;
+
+export const getClientRuntimeEnv = () =>
+  typeof window === "undefined"
+    ? null
+    : clientRuntimeSchema.parse(
+        window[globalVariableName as keyof typeof window]
+      );