+++ src/graphql/resolvers.ts
@@ -47,6 +47,8 @@ import { fetchOperatorInfo, search } from "src/rdf/search-queries";
 import { asElectricityCategory } from "src/domain/data";
 import { asNetworkLevel } from "src/domain/sunshine";
 import { last, sortBy } from "lodash";
+import { CoverageCacheManager } from "src/rdf/coverage-ratio";
+import { sparqlClient } from "src/rdf/sparql-client";
 
 const gfmSyntax = require("micromark-extension-gfm");
 const gfmHtml = require("micromark-extension-gfm/html");
@@ -66,6 +68,11 @@ const expectedCubeDimensions = [
   "https://energy.ld.admin.ch/elcom/electricityprice/dimension/product",
   "https://cube.link/observedBy",
 ];
+
+/**
+ * Under this threshold, observations are not returned
+ */
+const COVERAGE_RATIO_THRESHOLD = 0.25;
 
 const Query: QueryResolvers = {
   sunshineData: async (_parent, args, context) => {
@@ -216,9 +223,29 @@ const Query: QueryResolvers = {
     const operatorObservations = rawOperatorObservations.map((o) => ({
       __typename: "OperatorObservation",
       ...o,
-    }));
+    })) as ResolvedOperatorObservation[];
 
-    return operatorObservations as ResolvedOperatorObservation[];
+    const years = filters?.period;
+    if (years && observationFields && "coverageRatio" in observationFields) {
+      const defaultNetworkLevel = "NE7";
+      const coverageManager = new CoverageCacheManager(sparqlClient);
+      await coverageManager.prepare(years);
+      operatorObservations.forEach((x) => {
+        x.coverageRatio =
+          coverageManager.getCoverage(x, defaultNetworkLevel) ?? 1;
+        return x;
+      });
+    }
+
+    return operatorObservations.filter((o) => {
+      if (
+        o.coverageRatio !== undefined &&
+        o.coverageRatio < COVERAGE_RATIO_THRESHOLD
+      ) {
+        return false;
+      }
+      return true;
+    });
   },
   cantonMedianObservations: async (
     _,
@@ -276,16 +303,16 @@ const Query: QueryResolvers = {
         : [];
 
     const medianObservations = rawMedianObservations.map((x) => ({
-      __typename: "MedianObservation",
+      __typename: "CantonMedianObservation",
       ...x,
-    }));
+    })) as ResolvedCantonMedianObservation[];
 
-    return medianObservations as ResolvedCantonMedianObservation[];
+    return medianObservations;
   },
   swissMedianObservations: async (_, { locale, filters }, ctx, info) => {
-    let cantonCube;
+    let swissCube;
     try {
-      cantonCube = await getElectricityPriceSwissCube();
+      swissCube = await getElectricityPriceSwissCube();
     } catch (e: unknown) {
       const message = `${e instanceof Error ? e.message : e}`;
       console.error(message);
@@ -296,7 +323,7 @@ const Query: QueryResolvers = {
       });
     }
 
-    const cantonObservationsView = getView(cantonCube);
+    const swissObservationsView = getView(swissCube);
 
     // Look ahead to select proper dimensions for query
     const medianObservationFields = getResolverFields(
@@ -318,8 +345,8 @@ const Query: QueryResolvers = {
       medianDimensionKeys.length > 0
         ? await getElectricityPriceObservations(
             {
-              view: cantonObservationsView,
-              source: cantonCube.source,
+              view: swissObservationsView,
+              source: swissCube.source,
               isCantons: true,
               locale: locale ?? defaultLocale,
             },
@@ -332,10 +359,10 @@ const Query: QueryResolvers = {
 
     const medianObservations = rawMedianObservations.map((o) => ({
       ...o,
-      __typename: "MedianObservation",
-    }));
+      __typename: "SwissMedianObservation",
+    })) as ResolvedSwissMedianObservation[];
 
-    return medianObservations as ResolvedSwissMedianObservation[];
+    return medianObservations;
   },
   operators: async (_, { query, ids, locale }) => {
     const results = await search({