+++ CHANGELOG.md
@@ -9,15 +9,46 @@ and this project adheres to
 You can also check the
 [release page](https://github.com/visualize-admin/electricity-prices-switzerland/releases)
 
-# Unreleased
+# 2.11.3 (2025-08-26)
 
 - Feat
+
+  - Smaller font sizes for combobox and multi-combobox
+  - Button and IconButton sizes
+
+- Fix
+  - Active state of sidebar links for sunshine sidebar items
+
+# 2.11.1 (2025-08-26)
+
+- `CURRENT_PERIOD` is 2026
+
+# 2.11.0 (2025-08-26)
+
+- Fix
+  - `PUBLIC_URL` can be configured at runtime
+
+# 2.10.0 (2025-08-26)
+
+- Feat
+
   - Added OG tags to the app
   - Added custom tooltip to all button groups
 
 - Style
   - Changed Median Peer Group to correct color and shape
   - Introduced Median Total to line charts  
+  - Support for meteringrate and annualmeteringcost predicates
+    - Annual metering cost is shown on the map (CHF/year)
+    - Metering rate is derived from annual metering cost and shown in Rp./kWh to be consistent with other price components. It is displayed in the operator details page.
+  - Support for operator coverage ratio: The coverage ratio is used to filter out observations with a coverage ratio inferior to 25% and used to compute a weighted mean when a municipality has multiple operators.
+  - Improved the styling of the info banner fixed alignment
+  - 
+- Fix
+
+  - Minor UI styling issues
+  - Added custom tooltip to all button groups
+ 
 
 # 2.9.0